Add groupby aggregate tests, along with a few additons for cudf integration for SDFs
